feat: remove edit, update docs and refactor
